api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-storage==2.10.0
psycopg[binary]>=3.0
python-dotenv==1.0.0